Made changes to Update todo

+++ TrainingProject1/src/main/java/com/example/demo/model/RequirementSummarizationModel.java
@@ -14,7 +14,7 @@ public class RequirementSummarizationModel {
 	private int no_of_tasks;
 	private int no_of_task_completed;
 	private int no_of_task_notcompleted;
-	private float completionPercentage;
+	private double completionPercentage;
 	private List<TaskModel> reqTasks;
 	
 	public String getReq_Id() {
@@ -41,10 +41,10 @@ public class RequirementSummarizationModel {
 	public void setNo_of_task_notcompleted(int no_of_task_notcompleted) {
 		this.no_of_task_notcompleted = no_of_task_notcompleted;
 	}
-	public float getCompletionPercentage() {
+	public double getCompletionPercentage() {
 		return completionPercentage;
 	}
-	public void setCompletionPercentage(float completionPercentage) {
+	public void setCompletionPercentage(double completionPercentage) {
 		this.completionPercentage = completionPercentage;
 	}
 	public List<TaskModel> getReqTasks() {
@@ -53,9 +53,5 @@ public class RequirementSummarizationModel {
 	public void setReqTasks(List<TaskModel> reqTasks) {
 		this.reqTasks = reqTasks;
 	}
-	
-	
-	
-	
 	
 }
